Added a way to update the texture atlas to the current state of the shapes ( updates offsets and dest size ). A lot of minor fixes.

+++ src/ui/cuimenu.cpp
@@ -449,7 +449,7 @@ void cUIMenu::TrySelect( cUIControl * Ctrl, bool Up ) {
 			if ( Index != 0xFFFFFFFF ) {
 				if ( Up ) {
 					if ( Index > 0 ) {
-						for ( Uint32 i = Index - 1; i >= 0; i-- ) {
+						for ( Int32 i = (Int32)Index - 1; i >= 0; i-- ) {
 							if ( !mItems[i]->IsTypeOrInheritsFrom( UI_TYPE_SEPARATOR ) ) {
 								SetItemSelected( mItems[i] );
 								return;
